pyproject.toml
[tool.poetry]
package-mode = false

[project]
name = "leapscope"
version = "0.1.0"
description = "Professional-grade options analytics system for LEAPS"
authors = [
    {name = "User", email = "user@example.com"},
]
dependencies = [
    "yfinance>=0.2.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.10.0",
    "pyyaml>=6.0",
    "pydantic>=2.0.0",
    "scipy>=1.10.0",
    "streamlit>=1.30.0",
    "requests>=2.28.0",
    "pytest>=7.0.0",
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
